pyproject.toml
[tool.poetry]
package-mode = false
packages = [{include = "freelas", from = "src"}]


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.1.0"


[project]
name = "freelas"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["httpx (>=0.28.1,<0.29.0)", "lxml (>=5.4.0,<6.0.0)", "beautifulsoup4 (>=4.13.4,<5.0.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
